requirements.txt
# Core dependencies for Tic-Tac-Toe AI Agents
streamlit>=1.28.0
langchain>=0.1.0
langchain-experimental>=0.0.50
langchain-openai>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.6.0